+++ src/muz/spacer/spacer_util.h
@@ -121,6 +121,7 @@ void ground_expr(expr *e, expr_ref &out, app_ref_vector &vars);
 void mbqi_project(model &mdl, app_ref_vector &vars, expr_ref &fml);
 
 bool contains_selects(expr *fml, ast_manager &m);
+bool contains_defaults(expr *fml, ast_manager &m);
 void get_select_indices(expr *fml, app_ref_vector &indices);
 
 void find_decls(expr *fml, app_ref_vector &decls, std::string &prefix);