+++ src/qe/qsat.cpp
@@ -39,6 +39,7 @@ Notes:
 #include "qe/qe_mbp.h"
 #include "qe/qe.h"
 #include "ast/rewriter/label_rewriter.h"
+#include "util/params.h"
 
 namespace qe {
 
@@ -164,8 +165,9 @@ namespace qe {
         TRACE("qe_assumptions", model_v2_pp(tout, *mdl););
 
         expr_ref val(m);
-        for (unsigned j = 0; j < m_preds[level - 1].size(); ++j) {
-            app* p = m_preds[level - 1][j].get();            
+        for (unsigned i = 0; i <= level-1; ++i) {
+          for (unsigned j = 0; j < m_preds[i].size(); ++j) {
+            app* p = m_preds[i][j].get();            
             eval(p, val); 
             if (!m.inc())
                 return;
@@ -176,6 +178,7 @@ namespace qe {
                 SASSERT(m.is_true(val));
                 m_asms.push_back(p);
             }
+          }
         }
         asms.append(m_asms);
         
@@ -529,11 +532,14 @@ namespace qe {
         ast_manager& m;
         params_ref   m_params;
         ref<solver>  m_solver;
+
+        expr_ref m_last_assert;
         
     public:
         kernel(ast_manager& m):
             m(m),
-            m_solver(nullptr)
+            m_solver(nullptr),
+            m_last_assert(m)
         {
             m_params.set_bool("model", true);
             m_params.set_uint("relevancy", 0);
@@ -544,7 +550,7 @@ namespace qe {
         solver const& s() const { return *m_solver; }
 
         void init() {
-            m_solver = mk_smt_solver(m, m_params, symbol::null);
+           m_solver = mk_smt2_solver(m, m_params, symbol::null);
         }
         void collect_statistics(statistics & st) const {
             if (m_solver) 
@@ -561,7 +567,23 @@ namespace qe {
         void clear() {
             m_solver = nullptr;
         }
-        void assert_expr(expr* e) {
+
+      void assert_expr(expr *e) {
+        if (!m.is_true(e)) 
+          m_solver->assert_expr(e);
+      }
+        void assert_blocking_fml(expr* e) {
+          if (m.is_true(e)) return;
+          if (m_last_assert) {
+            if (e == m_last_assert) {
+              verbose_stream() << "Asserting this expression twice in a row:\n " << m_last_assert << "\n";
+              SASSERT(false);
+              std::exit(3);
+            }
+            
+          }
+          m_last_assert = e;
+
             m_solver->assert_expr(e);
         }
         
@@ -618,7 +640,9 @@ namespace qe {
         lbool check_sat() {        
             while (true) {
                 ++m_stats.m_num_rounds;
-                IF_VERBOSE(3, verbose_stream() << "(check-qsat level: " << m_level << " round: " << m_stats.m_num_rounds << ")\n";);
+                IF_VERBOSE(1, verbose_stream() << "(check-qsat level: " << m_level << " round: " << m_stats.m_num_rounds << ")\n";);
+                TRACE("qe",
+                      tout << "level: " << m_level << " round: " << m_stats.m_num_rounds << "\n");
                 check_cancel();
                 expr_ref_vector asms(m_asms);
                 m_pred_abs.get_assumptions(m_model.get(), asms);
@@ -951,7 +975,8 @@ namespace qe {
             }
             else {
                 fml = m_pred_abs.mk_abstract(fml);
-                get_kernel(m_level).assert_expr(fml);
+                TRACE("qe_block", tout << "Blocking fml at level: " << m_level << "\n" << fml << "\n";);
+                get_kernel(m_level).assert_blocking_fml(fml);
             }
             SASSERT(!m_model.get());
             return true;
@@ -1235,8 +1260,11 @@ namespace qe {
             m_value(nullptr),
             m_was_sat(false),
             m_gt(m)
-        {
-        }
+            {
+                params_ref q = params_ref();
+                q.set_bool("use_qel", false);
+                m_mbp.updt_params(q);
+            }
         
         ~qsat() override {
             clear();